refactor: move response data type to types/App
