+++ src/pages/BlockList/index.tsx
@@ -22,8 +22,8 @@ import BlockRewardIcon from '../../assets/block_reward_white.png'
 import MinerIcon from '../../assets/miner.png'
 import TimestampIcon from '../../assets/timestamp.png'
 import { fetchBlockList } from '../../service/fetcher'
-import { BlockWrapper } from '../../service/response/Block'
-import { Response } from '../../service/response/Response'
+import { Block, Wrapper } from '../../types/App/index'
+import { Response } from '../../types/App/Response'
 import { shannonToCkb } from '../../utils/util'
 import { parsePageNumber } from '../../utils/string'
 import { CachedKeys } from '../../utils/const'
@@ -61,7 +61,7 @@ const reducer = (state: any, action: any) => {
 
 const getBlocks = (page: number, size: number, dispatch: any) => {
   fetchBlockList(page, size).then(response => {
-    const { data, meta } = response as Response<BlockWrapper[]>
+    const { data, meta } = response as Response<Wrapper<Block>[]>
     if (meta) {
       dispatch({
         type: Actions.total,
@@ -83,7 +83,7 @@ const getBlocks = (page: number, size: number, dispatch: any) => {
 }
 
 const initialState = {
-  blocks: [] as BlockWrapper[],
+  blocks: [] as Wrapper<Block>[],
   total: 1,
 }
 
@@ -100,7 +100,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps>) => {
   const [state, dispatch] = useReducer(reducer, initialState)
 
   useEffect(() => {
-    const cachedBlocks = fetchCachedData<BlockWrapper[]>(CachedKeys.BlockList)
+    const cachedBlocks = fetchCachedData<Wrapper<Block>[]>(CachedKeys.BlockList)
     if (cachedBlocks) {
       dispatch({
         type: Actions.blocks,
@@ -146,7 +146,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps>) => {
               <TableTitleItem image={TimestampIcon} title={t('home.time')} />
             </TableTitleRow>
             {state.blocks &&
-              state.blocks.map((data: BlockWrapper) => {
+              state.blocks.map((data: Wrapper<Block>) => {
                 return (
                   data && (
                     <TableContentRow key={data.attributes.block_hash}>
