--- src/service/response/Data.ts
@@ -1,9 +1,0 @@
-export interface Data {
-  data: string
-}
-
-export interface DataWrapper {
-  id: number
-  type: string
-  attributes: Data
-}
