--- src/service/response/Script.ts
@@ -1,10 +1,0 @@
-export interface Script {
-  code_hash: string
-  args: string[]
-}
-
-export interface ScriptWrapper {
-  id: number
-  type: string
-  attributes: Script
-}
