+++ src/utils/const.ts
@@ -10,3 +10,13 @@ export const CachedKeys = {
   Blocks: 'blocks',
   Statistics: 'statistics',
 }
+
+export enum TransactionFeeStatus {
+  calculating = 'calculating',
+  calculated = 'calculated',
+}
+
+export enum RewardStatus {
+  pending = 'pending',
+  issued = 'issued',
+}
